docs: 增加xcode-keybinding.sh的说明

+++ xcode-keybinding.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 ##author: zhoujie<zhoujie_903@163.com>
-##这个脚本的作用是：自定义xcode的键盘快捷键，Xcode 10有效
+##这个脚本的作用是：自定义Xcode的键盘快捷键，在Xcode 10上测试有效
 
 # /Applications/Xcode.app/Contents/Frameworks/IDEKit.framework/Resources/IDETextKeyBindingSet.plist
 # /Users/zhoujie/Library/Developer/Xcode/UserData/KeyBindings
@@ -8,9 +8,12 @@
 PLIST="IDETextKeyBindingSet.plist"
 KeyBindingSetPlist="/Applications/Xcode.app/Contents/Frameworks/IDEKit.framework/Resources/$PLIST"
 DST="~/Desktop"
-# echo "从Xcode复制${PLIST}到桌面"
-# cp $KeyBindingSetPlist ~/Desktop
 
+# 1
+echo "从Xcode复制${PLIST}到桌面"
+cp $KeyBindingSetPlist ~/Desktop
+
+# 2
 # 编辑文件
 	# <key>My Custom Actions</key>
 	# <dict>
@@ -20,10 +23,19 @@ DST="~/Desktop"
 	# 	<string>moveUp:, moveToEndOfLine:, insertNewline:</string>
 	# </dict>
 
+echo "编辑好${PLIST}后" 
+# -p prompt
+# If the -r option is given, this signifies `raw' input, and backslash escaping is disabled.    
+read -r -p "点击Y/y键将从桌面${PLIST}复制到Xcode,其它键退出：" confirmation
+if [ "$confirmation" != y ] && [ "$confirmation" != Y ]; then
+  exit
+fi
+
+# 3
 echo "从桌面${PLIST}复制到Xcode"
 sudo mv ~/Desktop/$PLIST $KeyBindingSetPlist
 
-open /Applications/Xcode.app/Contents/Frameworks/IDEKit.framework/Resources/
+# open /Applications/Xcode.app/Contents/Frameworks/IDEKit.framework/Resources/
 
 
 # 参考：
